# docs/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: Brix Nutritional App API
  version: 2.0.0
  description: |
    REST API для Brix Nutritional App (Mobile + Admin Web Panel).
    
    ## Компоненты системы
    
    ### 1. Mobile App (Flutter)
    Мобильное приложение для пользователей (iOS/Android).
    Использует все endpoints из категорий: Auth, Users, Home, Meal Plans, Recipes, Diary, AI Chat, Lab Tests, Courses, Blog, Notifications, Subscriptions.
    
    ### 2. Admin Web Panel (Next.js)
    Веб-панель для администраторов и контент-менеджеров.
    Использует admin-специфичные endpoints и endpoints для управления контентом (Courses, Lessons, Recipes, Meal Plans, Users, Blog, Subscriptions, Analytics).
    
    ### 3. Backend API (Strapi)
    Сервер на Node.js + Strapi CMS с кастомными контроллерами.
    
    ---
    
    ## Авторизация
    
    ### Для Mobile App:
    Большинство endpoints требуют JWT токен в заголовке:
    ```
    Authorization: Bearer <token>
    ```
    
    ### Для Admin Web Panel:
    Admin endpoints требуют Strapi admin JWT токен:
    ```
    Authorization: Bearer <admin-token>
    ```
    Токен получается через Strapi `/admin/login` endpoint.
    
    ---
    
    ## Rate Limiting
    - Auth endpoints: 5 запросов / 15 минут
    - Mobile App endpoints: 100 запросов / минуту
    - Admin endpoints: 1000 запросов / минуту
    
  contact:
    name: API Support
    email: support@brix-nutrition.com
servers:
  - url: http://localhost:1337/api
    description: Development
...
paths:
...
  /recipes/{id}:
    get:
      tags: [Recipes]
      summary: Получить рецепт
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Рецепт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
  /recipes/{id}/alternatives:
    get:
      tags: [Recipes]
      summary: Получить альтернативные рецепты
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Альтернативы
          content:
            application/json:
              schema:
                type: object
                properties:
                  alternatives:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recipe'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        prepTime:
          type: integer
          description: Минуты
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fats:
          type: number
        instructions:
          type: array
          items:
            type: string
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              amount:
                type: number
              unit:
                type: string
        tags:
          type: array
          items:
            type: string
...